games/factorio/requirements.txt
factorio-rcon-py==2.0.1
paho-mqtt==1.6.1
python-dotenv==0.21.0